requirements.txt
selenium
psycopg2-binary
webdriver-manager
python-dotenv
fastapi
sqlalchemy
uvicorn
pydantic
alembic
openai
langchain
sentence-transformers